+++ NEWS.md
@@ -1,3 +1,18 @@
+# jsTreeR 2.4.0 (2023-11-07)
+
+- New option `checkWithText`; it is a `TRUE`/`FALSE` option about the behavior 
+of the checkboxes: it controls whether the checkboxes must be (un)checked when
+clicking on the text of a node (thanks @ismirsehregal for the feature request).
+
+- The Shiny input value `ID_selected_tree` (see `?jstree-shiny`) has been 
+renamed to `ID_checked_tree`, and there are new Shiny values provided when 
+the checkboxes are enabled. Their description is given in `?jstree-shiny`.
+
+- New feature for grids: when `search=TRUE`, there is now one search box for 
+each column of the grid (thanks @ismirsehregal for the feature request and for 
+the help).
+
+
 # jsTreeR 2.3.3 (2023-10-10)
 
 - Upgraded jsTree to 3.3.16.